+++ DmxLedPanel/DmxLedPanel/Util/Utils.cs
@@ -13,18 +13,18 @@ namespace DmxLedPanel
         public static readonly int DMX_VALUE_MAX = 255;
         public static readonly int DMX_VALUE_MIN = 0;
 
-        public static void checkDmxValueRange(int value) {
+        public static void checkDmxValueRange(byte value) {
             if (value > DMX_VALUE_MAX || value < DMX_VALUE_MIN) {
                 throw new ArgumentOutOfRangeException(value + " is out of dmx value range (" + DMX_VALUE_MIN + " to " + DMX_VALUE_MAX + ")");
             }
         }
 
-        public static void checkDmxValueRange(int[] values) {
-            foreach (int val in values) {
+        public static void checkDmxValueRange(byte[] values) {
+            foreach (byte val in values) {
                 checkDmxValueRange(val);
             }
         }
-
+        
         public static T[] GetSubArray<T>(T[] data, int index, int length) {
             T[] result = new T[length];
             Array.Copy(data, index, result, 0, length);
